Rename LimitedConnectionPool to StreamLimitingConnectionPool
Alias added for BC.
